Moved mapper to persistence package
