+++ src/main/java/com/globallogic/amcr/persistence/mapper/casestudies/CaseStudyMapper.java
@@ -1,4 +1,4 @@
-package com.globallogic.amcr.mapper.casestudies;
+package com.globallogic.amcr.persistence.mapper.casestudies;
 
 import com.globallogic.amcr.persistence.model.casestudies.CaseStudy;
 import com.globallogic.amcr.persistence.model.casestudies.CaseStudyOverview;
