+++ src/main/java/com/globallogic/amcr/persistence/mapper/contactcomponent/FileMapper.java
@@ -1,4 +1,4 @@
-package com.globallogic.amcr.mapper.contactcomponent;
+package com.globallogic.amcr.persistence.mapper.contactcomponent;
 
 import com.globallogic.amcr.persistence.model.contactcomponent.Attachment;
 import com.globallogic.amcr.persistence.payload.contactcomponent.AttachmentMetadata;
